hw/xbox/mcpx/apu/vp: drop float->double promotion in LPF + SVF
Hypothesis: setup_svf used sqrt() + double literals, and the per-
sample LPF post-clamp used fmin/fmax + 1.0 (double). Both force an
ARM64 float->double->float round-trip instead of a single-precision
fsqrt on S-reg / fminnm/fmaxnm on S-reg. Each per-sample clamp runs
32x2 times per LPF-active voice per frame.

Fix: sqrtf, fminf, fmaxf, and 1.0f/0.5f/0.01f float literals. Precision
unchanged at the 8-bit voice quantization level.

Test: ReadLints clean.

Trade-off: trivial. No conflict with reverted-experiment entries.
Made-with: Cursor

+++ hw/xbox/mcpx/apu/vp/vp.c
@@ -1658,7 +1658,7 @@ static void voice_process(MCPXAPUState *d,
             setup_svf(filter, fc_f, q_f, F_LP);
             for (int i = 0; i < NUM_SAMPLES_PER_FRAME; i++) {
                 samples[i][ch] = run_svf(filter, samples[i][ch]);
-                samples[i][ch] = fmin(fmax(samples[i][ch], -1.0), 1.0);
+                samples[i][ch] = fminf(fmaxf(samples[i][ch], -1.0f), 1.0f);
             }
         }
     }
